--- view/adminhtml/templates/config/allowed_origin.phtml
@@ -1,68 +1,0 @@
-<?php
-/** @var \Adyen\Payment\Model\Config\Adminhtml\AllowedOriginField $block */
-echo $block->getButtonHtml()
-?>
-<span class="adyen-required-settings-config-admin" id="allowed_origin_progress">
-    <img class="processing" hidden="hidden" alt="Configuring" style="margin:0 5px" src="<?php echo $block->getViewFileUrl('images/process_spinner.gif') ?>"/>
-    <img class="configured" hidden="hidden" alt="Configured" style="margin:-3px 5px" src="<?php echo $block->getViewFileUrl('images/rule_component_apply.gif') ?>"/>
-    <span id="adyen_allowed_origin_configured_message"></span>
-    <div id="adyen_allowed_origin_configuration_errors" class="message-system-inner">
-        <div class="message message-warning"></div>
-    </div>
-</span>
-<script>
-    require([
-            'jquery',
-            'prototype'
-        ],
-        function(jQuery)
-        {
-            jQuery('#adyen_allowed_origin_configuration_errors').hide();
-            let progressSpan = jQuery('#allowed_origin_progress');
-
-            jQuery('#adyen_configure_allowed_origin').click(function () {
-                const storeId = '<?=$block->getStoreId()?>';
-
-                new Ajax.Request('<?php echo $block->getAjaxUrl() ?>', {
-                    parameters: {storeId},
-                    type: "POST",
-                    loaderArea:     false,
-                    asynchronous:   true,
-                    onCreate: function() {
-                        progressSpan.find('.configured').hide();
-                        progressSpan.find('.processing').show();
-                        jQuery('#adyen_allowed_origin_configured_message').text('');
-                        jQuery("#adyen_configure_allowed_origin").prop('disabled', true);
-                    },
-                    onSuccess: function (response) {
-                        //add html response
-                        progressSpan.find('.processing').hide();
-                        let resultText = '';
-                        if (response.status > 200) {
-                            resultText = response.statusText;
-                        } else if (response.responseJSON === null) {
-                            resultText = "Could not retrieve the response";
-                        } else {
-                            const allowedOriginInput = document.querySelector(
-                                "div.adyen_required_config_settings input[name*='allowed_origin']");
-                           allowedOriginInput.value = response.responseJSON.originUrl;
-
-                            resultText = 'Completed';
-                            progressSpan.find('.configured').show();
-                        }
-                        jQuery('#adyen_allowed_origin_configured_message').text(resultText);
-                    },
-                    onFailure: function(response)
-                    {
-                        progressSpan.find('.processing').hide();
-                        jQuery('#adyen_allowed_origin_configured_message').text("Failed").show();
-                    },
-                    onComplete: function()
-                    {
-                        jQuery("#adyen_configure_allowed_origin").prop('disabled', false);
-                    }
-                });
-            });
-
-        });
-</script>
