+++ view/adminhtml/templates/config/required_settings.phtml
@@ -1,3 +1,5 @@
+<?php /** @var \Adyen\Payment\Model\Config\Adminhtml\MerchantAccounts $block */
+echo $block->getButtonHtml() ?>
 <script>
     require([
             'jquery',
@@ -19,8 +21,9 @@
                     xapi_key = document.querySelector(
                         "div.adyen_required_config_settings input[name*='api_key_live']").value;
                 }
+                const storeId = '<?=$block->getStoreId()?>';
                 new Ajax.Request('<?php echo $block->getAjaxUrl() ?>', {
-                    parameters: {xapikey: xapi_key, demoMode: demo_mode},
+                    parameters: {xapikey: xapi_key, demoMode: demo_mode, storeId},
                     type: "POST",
                     loaderArea:     false,
                     asynchronous:   true,
@@ -62,6 +65,9 @@
                                     "div.adyen_required_config_settings input[name*='client_key_live']");
                                 clientKeyIdLive.value = clientkey;
                             }
+                            const allowedOriginInput = document.querySelector(
+                                "div.adyen_required_config_settings input[name*='allowed_origin']");
+                            allowedOriginInput.value = response.responseJSON.originUrl;
                             resultText = 'Completed';
                             progressSpan.find('.configured').show();
                         }
@@ -82,7 +88,6 @@
         });
 </script>
 
-<?php echo $block->getButtonHtml() ?>
 <span class="adyen-required-settings-config-admin" id="progress">
     <img class="processing" hidden="hidden" alt="Configuring" style="margin:0 5px" src="<?php echo $block->getViewFileUrl('images/process_spinner.gif') ?>"/>
     <img class="configured" hidden="hidden" alt="Configured" style="margin:-3px 5px" src="<?php echo $block->getViewFileUrl('images/rule_component_apply.gif') ?>"/>
@@ -93,5 +98,5 @@
 </span>
 <!--if we want to add a note under the button-->
 <p class="note">
-    <span><?php echo __('Retrieves the associated merchant accounts and client key'); ?></span>
+    <span><?php echo __('Retrieves the associated merchant accounts, client key and prefills the allowed origin'); ?></span>
 </p>
